ADD_MAKEFILE_FINISH_EX03

+++ 03_cpp/01_cpp/ex01/src/zombieHorde.cpp
@@ -6,7 +6,7 @@
 /*   By: jfranco <jfranco@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 17:13:39 by jfranco           #+#    #+#             */
-/*   Updated: 2025/04/11 17:56:22 by jfranco          ###   ########.fr       */
+/*   Updated: 2025/05/02 15:56:00 by jfranco          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ Zombie*	zombieHorde(size_t i, std::string name)
 {
 	Zombie* nZombie = new Zombie[i];
 	i--;
-	while (i != -1)
+	while ((int)i != -1)
 	{
 		nZombie[i].set_name(name);
 		i--;
